@@ -6,14 +6,13 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^13.1.4",
-    "axios": "^0.21.1",
-    "create-react-app": "^4.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "axios": "^1.2.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-html5-camera-photo": "^1.5.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-toastify": "^7.0.3"
+    "react-router-dom": "^6.6.2",
+    "react-scripts": "5.0.1",
+    "react-toastify": "^9.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
