@@ -1,7 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import { toast } from 'react-toastify'
+import React, { Fragment } from 'react';
+import ReactDOM from 'react-dom/client';
+import { ToastContainer } from 'react-toastify'
 
 import 'react-toastify/dist/ReactToastify.css';
 import 'css/bootstrap.min.css';
@@ -9,15 +8,10 @@ import 'css/style.css';
 
 import App from 'app';
 
-toast.configure({
-	autoClose: 3000,
-	draggable: false,
-	hideProgressBar: true
-});
-
-ReactDOM.render(
-	<BrowserRouter>
-		<App />
-	</BrowserRouter>,
-	document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+	<Fragment>
+    	<App />
+		<ToastContainer theme="dark" position="top-right" autoClose={3000} hideProgressBar={false} newestOnTop={false} closeOnClick />
+	</Fragment>
 );
