@@ -1,5 +1,5 @@
 import React, { Suspense, lazy } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import Loading from 'components/common/loading';
 import ErrorBoundary from 'components/common/error_boundary';
@@ -13,34 +13,29 @@ const Camera = lazy(() => import('components/camera/index'));
 const Upload = lazy(() => import('components/upload/index'));
 
 /**************************************************************************************/
-export default class App extends React.Component {
+function App() {	
+	return (
+		<BrowserRouter>
+			<Header />
 
-	/********************************************/
-	render(){
-		return(
-			<Router>
-				<Header />
-
-				<div className="content">
-					<ErrorBoundary>
-						<Suspense fallback={<Loading />}>
-							<Switch>
-								<Route exact path="/" component={Homepage} />
-								<Route path="/recipe/:id" component={Recipe} />
-								<Route exact path="/tags" component={Tags} />
-								<Route exact path="/search/" component={Search} />
-								<Route exact path="/camera" component={Camera} />
-								<Route exact path="/upload" component={Upload} />
-							</Switch>
-						</Suspense>
-					</ErrorBoundary>
-				</div>
-				
-			</Router>
-		);
-	}
-	/********************************************/
-
+			<div className="content">
+				<ErrorBoundary>
+					<Suspense fallback={<Loading />}>
+						<Routes>
+							<Route exact path="/" element={<Homepage />} />
+							<Route path="/recipe/:id" element={<Recipe />} />
+							<Route exact path="/tags" element={<Tags />} />
+							<Route exact path="/search/" element={<Search />} />
+							<Route exact path="/camera" element={<Camera />} />
+							<Route exact path="/upload" element={<Upload />} />
+						</Routes>
+					</Suspense>
+				</ErrorBoundary>
+			</div>
+			
+		</BrowserRouter>
+	);
 }
 /**************************************************************************************/
 
+export default App;
