@@ -1,39 +1,34 @@
-import React, { Fragment } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 
 import API from 'api/api';
-import refresh from 'images/refresh.png';
+import RefreshIcon from 'images/refresh.png';
 import Recipe from 'components/common/recipe';
 
 /**************************************************************************************/
-export default class Homepage extends React.Component {
+function Homepage(){
 
-	/****************************************/
-	constructor(props) {
-		super(props);
-
-		this.state = {
-			recipe: {},
-		}
-
-		this.refresh = this.refresh.bind(this);
-	}
+    /****************************************/
+    const [state, setValues] = useState({
+        recipe: {}
+    });
 	/****************************************/
 
 
 	/****************************************/
-	componentDidMount(){
-		this.refresh();
-	}
+	useEffect(() => {
+		refresh();
+	},[]);	
 	/****************************************/
-
+	
 
 	/****************************************/
-	refresh(){
+	const refresh = () =>{
+		console.log('refresh');
 		document.getElementById('spinner-holder').style.display = 'block';
 		API.getBookmark().then((data) => {
 			document.getElementById('spinner-holder').style.display = 'none';
 
-			this.setState({
+			setValues({
 				recipe: data
 			});
 		});
@@ -42,31 +37,30 @@ export default class Homepage extends React.Component {
 
 
 	/****************************************/
-	render(){
-		return(
-			<Fragment>
-				<div className="row button-row">
-					<div className="col-12">
-						<button className="btn refresh" onClick={this.refresh}><img src={refresh} alt="refresh" /></button>
-					</div>
+	return(
+		<Fragment>
+			<div className="row button-row">
+				<div className="col-12">
+					<button className="btn refresh" onClick={refresh}><img src={RefreshIcon} alt="refresh" /></button>
 				</div>
+			</div>
 
-				{typeof this.state.recipe.id !== 'undefined' &&
-					<Recipe recipe={this.state.recipe} />
-				}
+			{typeof state.recipe.id !== 'undefined' &&
+				<Recipe recipe={state.recipe} />
+			}
 
-				{typeof this.state.recipe.id === 'undefined' &&
-					<Fragment>
-						<div className="row">
-							<div className="col-12 text-center">Nothing Bookmarked :-(</div>
-						</div>			
-					</Fragment>
-				}
-			</Fragment>
-		)
-	}
+			{typeof state.recipe.id === 'undefined' &&
+				<Fragment>
+					<div className="row">
+						<div className="col-12 text-center">Nothing Bookmarked :-(</div>
+					</div>			
+				</Fragment>
+			}
+		</Fragment>
+	);
 	/****************************************/
 
 }
 /**************************************************************************************/
 
+export default Homepage;
